refactor:getUserNicknameByEmail함수 모듈화
